pro30  답변형게시판 만들기

+++ pro30/src/main/java/com/myspring/pro30/board/controller/BoardControllerImpl.java
@@ -2,6 +2,7 @@ package com.myspring.pro30.board.controller;
 
 import com.myspring.pro30.board.service.BoardService;
 import com.myspring.pro30.board.vo.ArticleVO;
+import com.myspring.pro30.board.vo.ImageVO;
 import com.myspring.pro30.member.vo.MemberVO;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
@@ -50,49 +51,57 @@ public class BoardControllerImpl implements BoardController {
         return mav;
     }
 
+
     @Override
     @RequestMapping(value = "/board/addNewArticle.do", method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity addNewArticle(MultipartHttpServletRequest multipartRequest,
-                                        HttpServletResponse response) throws Exception {
-        response.setContentType("text/html; charset=utf-8");
+    public ResponseEntity addNewArticle(MultipartHttpServletRequest multipartRequest, HttpServletResponse response) throws Exception {
         multipartRequest.setCharacterEncoding("utf-8");
-        Map<String, Object> articleMap = new HashMap<String, Object>();
+        String imageFileName = null;
+
+        Map articleMap = new HashMap();
         Enumeration enu = multipartRequest.getParameterNames();
         while (enu.hasMoreElements()) {
             String name = (String) enu.nextElement();
-
             String value = multipartRequest.getParameter(name);
-
-            System.out.println("name = " + name);
-            System.out.println("value = " + value);
-
-            name = new String(name.getBytes("8859_1"), "utf-8");
-            value = new String(value.getBytes("8859_1"), "utf-8");
-            System.out.println("name = " + name);
-            System.out.println("value = " + value);
             articleMap.put(name, value);
         }
 
-        String imageFileName = upload(multipartRequest);
+        //로그인 시 세션에 저장된 회원 정보에서 글쓴이 아이디를 얻어와서 Map에 저장합니다.
         HttpSession session = multipartRequest.getSession();
         MemberVO memberVO = (MemberVO) session.getAttribute("member");
         String id = memberVO.getId();
-        articleMap.put("parentNO", 0);
         articleMap.put("id", id);
-        articleMap.put("imageFileName", imageFileName);
 
+
+        List<String> fileList = upload(multipartRequest);
+        List<ImageVO> imageFileList = new ArrayList<ImageVO>();
+        if (fileList != null && fileList.size() != 0) {
+            for (String fileName : fileList) {
+                ImageVO imageVO = new ImageVO();
+                imageVO.setImageFileName(fileName);
+                imageFileList.add(imageVO);
+            }
+            articleMap.put("imageFileList", imageFileList);
+        }
         String message;
         ResponseEntity resEnt = null;
         HttpHeaders responseHeaders = new HttpHeaders();
         responseHeaders.add("Content-Type", "text/html; charset=utf-8");
         try {
             int articleNO = boardService.addNewArticle(articleMap);
-            if (imageFileName != null && imageFileName.length() != 0) {
-                File srcFile = new
-                        File(ARTICLE_IMAGE_REPO + "/" + "temp" + "/" + imageFileName);
-                File destDir = new File(ARTICLE_IMAGE_REPO + "/" + articleNO);
-                FileUtils.moveFileToDirectory(srcFile, destDir, true);
+            if (imageFileList != null && imageFileList.size() != 0) {
+                for (ImageVO imageVO : imageFileList) {
+                    imageFileName = imageVO.getImageFileName();
+                    imageFileName = new String(imageFileName.getBytes("UTF-8"));
+                    System.out.println("imageFileName = " + imageFileName);
+                    File srcFile = new File(ARTICLE_IMAGE_REPO + "/" + "temp" + "/" + imageFileName);
+                    System.out.println("srcFile.getName() = " + srcFile.getName());
+                    File destDir = new File(ARTICLE_IMAGE_REPO + "/" + articleNO);
+                    System.out.println("destDir.getName() = " + destDir.getName());
+                    destDir.mkdirs();
+                    FileUtils.moveFileToDirectory(srcFile, destDir, true);
+                }
             }
 
             message = "<script>";
@@ -100,9 +109,17 @@ public class BoardControllerImpl implements BoardController {
             message += " location.href='" + multipartRequest.getContextPath() + "/board/listArticles.do'; ";
             message += " </script>";
             resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+
+
         } catch (Exception e) {
-            File srcFile = new File(ARTICLE_IMAGE_REPO + "/" + "temp" + "/" + imageFileName);
-            srcFile.delete();
+            if (imageFileList != null && imageFileList.size() != 0) {
+                for (ImageVO imageVO : imageFileList) {
+                    imageFileName = imageVO.getImageFileName();
+                    File srcFile = new File(ARTICLE_IMAGE_REPO + "/" + "temp" + "/" + imageFileName);
+                    srcFile.delete();
+                }
+            }
+
 
             message = " <script>";
             message += " alert('오류가 발생했습니다. 다시 시도해 주세요');');";
@@ -122,6 +139,7 @@ public class BoardControllerImpl implements BoardController {
         return mav;
     }
 
+/*
     private String upload(MultipartHttpServletRequest multipartRequest) throws Exception {
         String imageFileName = null;
         Iterator<String> fileNames = multipartRequest.getFileNames();
@@ -144,16 +162,41 @@ public class BoardControllerImpl implements BoardController {
         }
         return imageFileName;
     }
+*/
+
+    private List<String> upload(MultipartHttpServletRequest multipartRequest) throws Exception {
+        List<String> fileList = new ArrayList<String>();
+        Iterator<String> fileNames = multipartRequest.getFileNames();
+        while (fileNames.hasNext()) {
+            String fileName = fileNames.next();
+            MultipartFile mFile = multipartRequest.getFile(fileName);
+            String originalFileName = mFile.getOriginalFilename();
+            fileList.add(originalFileName);
+            File file = new File(ARTICLE_IMAGE_REPO + "/" + fileName);
+            if (mFile.getSize() != 0) { //File Null Check
+                if (!file.exists()) { //경로상에 파일이 존재하지 않을 경우
+                    if (file.getParentFile().mkdirs()) { //경로에 해당하는 디렉토리들을 생성
+                        file.createNewFile(); //이후 파일 생성
+                    }
+                }
+                mFile.transferTo(new File(ARTICLE_IMAGE_REPO + "/" + "temp" + "/" + originalFileName)); //임시로 저장된 multipartFile을 실제 파일로 전송
+            }
+        }
+        return fileList;
+    }
+
 
     //한개의 이미지 보여주기
     @RequestMapping(value = "/board/viewArticle.do", method = RequestMethod.GET)
     public ModelAndView viewArticle(@RequestParam("articleNO") int articleNO,
                                     HttpServletRequest request, HttpServletResponse response) throws Exception {
         String viewName = (String) request.getAttribute("viewName");
-        articleVO = boardService.viewArticle(articleNO);
+//        articleVO = boardService.viewArticle(articleNO);
+        Map articleMap = boardService.viewArticle(articleNO);
         ModelAndView mav = new ModelAndView();
         mav.setViewName(viewName);
-        mav.addObject("article", articleVO);
+//        mav.addObject("article", articleVO);
+        mav.addObject("articleMap", articleMap);
         return mav;
     }
 
@@ -171,7 +214,8 @@ public class BoardControllerImpl implements BoardController {
             articleMap.put(name, value);
         }
 
-        String imageFileName = upload(multipartRequest);
+//        String imageFileName = upload(multipartRequest);
+        String imageFileName = null;
         HttpSession session = multipartRequest.getSession();
         MemberVO memberVO = (MemberVO) session.getAttribute("member");
         String id = memberVO.getId();
